Refactor app factory to separate module
Moved the Flask application factory from app/__init__.py to a new app/factory.py module. Updated imports in run.py and tests/conftest.py accordingly. Cleaned up __init__.py files in app, routes, and services packages to serve as package markers and updated README to reflect the new structure.
