+++ app/services/__init__.py
@@ -1,7 +1,1 @@
-"""Services package initialization."""
-
-from app.services.dividend_service import DividendService
-from app.services.investment_service import InvestmentService
-from app.services.portfolio_service import PortfolioService
-
-__all__ = ["InvestmentService", "DividendService", "PortfolioService"]
+"""Services package."""
