+++ tests/conftest.py
@@ -4,8 +4,8 @@ from typing import Generator
 
 import pytest
 
-from app import create_app
 from app.extensions import db
+from app.factory import create_app
 from app.models import Dividend, Investment
